Create MEMcoin.pl
This is the foundation of our code that will be elaborated on as we get closer to the deadline of the project. We are currently working on a regex that will allow us to find a formatted string that is close(in distance)to the wallet password in the memory dump.